+++ test/source/test_set_partition.cpp
@@ -3,7 +3,7 @@
 #include <ecgen/set_partition.hpp>
 
 TEST_CASE("set partition") {
-    int cnt = 1;
+    size_t cnt = 1;
     for ([[maybe_unused]] auto i : ecgen::set_partition_gen(5, 3)) {
         ++cnt;
     }
